backend/requirements.txt
flask
flask-cors
supabase
google-generativeai
python-decouple
PyPDF2
python-docx
